refactor(types): extract comment tests to sibling file
